Splitted DiagnosisKeyBatchService in Transactional and non Transactional class.
fixed Sonar Bug: https://sonarcloud.io/project/issues?id=eu-federation-gateway-service_efgs-federation-gateway&issues=AXRZDd9204CUos3cXweC&open=AXRZDd9204CUos3cXweC
